+++ springql-core/src/stream_engine/autonomous_executor/task_executor/source_worker_pool.rs
@@ -28,7 +28,7 @@ pub(super) struct SourceWorkerPool {
 
 impl SourceWorkerPool {
     pub(super) fn new(
-        n_worker_threads: usize,
+        n_worker_threads: u16,
         event_queue: Arc<EventQueue>,
         task_executor_lock: Arc<TaskExecutorLock>,
         repos: Arc<Repositories>,
